Extract Rivers Source/Layer to a component
